% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalancedExtensionSolver.R
\name{BalancedExtensionSolver}
\alias{BalancedExtensionSolver}
\title{Unbalanced Transport Solver}
\usage{
BalancedExtensionSolver(
  supplyList,
  demandList,
  costMethod,
  costExp,
  costP,
  wfr = FALSE,
  Monge = FALSE,
  costMatrix = NULL
)
}
\arguments{
\item{supplyList}{A list containing the information about the supply. The first element hast to be the distribution followed by
 a vector specifying the cost for destruction of mass at each supply point. If no cost matrix is provided, the third list element has to be 
the positions of the supply points. This can either be a vector or a matrix where each row gives the coordinates for one point.}

\item{demandList}{A list similar the the supplyList.}

\item{wfr}{(optional) Computes the cost matrix needed for the Wasserstein-Fisher-Rao distance \eqn{c(x,y) = -\log(\cos^2_+(d(x,y)))}{c(x,y) = -log(cos_+(d(x,y)²))}.
The default value is "false".}

\item{Monge}{(optional) Set to TRUE to use the Monge algorithm.}

\item{costMatrix}{(optional) Instead of having the algorithm compute the cost matrix, a custom cost matrix can be passed to the algorithm.}

\item{method}{(optional) Determines the method that is used to compute the cost matrix.
\itemize{
\item "euclidean"
\item "minkowski"
\item "maximum" 
}
The default value is "euclidean".}

\item{exp}{(optional) The exponent that is applied to the cost matrix. Can be used to compute quadratic cost. The default value is 1.}

\item{p}{(optional) Parameter for the minkwski cost function. Can be omitted if either "euclidean" or "maximum" is used. The default value is 2.}
}
\description{
This function solves unbalanced optimal transport problems by extending them to a balanced problem and applying the revised simplex algorithm.
}
\details{
The unbalanced optimal transport problem \eqn{<C,r> + \sum_i p_i(\alpha_i - \sum_j r_{ij}) + \sum_j q_j(\beta-\sum_i r{ij})}{
<C,r> + sum_i(p_i (a-sum_j r_ij)) + sum_j(q_j (b-sum_i r_ij))} with supply and demand measure \eqn{\alpha}{a} and \eqn{\beta}{b},
cost matrix C, transport plan \eqn{r} and construction and destruction costs \eqn{p} and \eqn{q} is transformed to a balanced transport 
problem by extending the cost matrix and supply and demand vectors. 

The resulting transport problem is then solved using the reverse simplex algorithm provided by the \code{\link[transport]{transport}} package.

If the cost matrix fulfils the Monge property, a algorithm designed take advantage of this special structure can be used as well. However, the 
Monge algorithm is not as efficient as the revised simplex algorithm.


\insertRef{Guittet2002}{unbalancedTransport}
}
