% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmMonge.R
\name{checkMongeProperty}
\alias{checkMongeProperty}
\title{checkMongeProperty}
\usage{
checkMongeProperty(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
'0' if the matrix does not fulfill the Monge property, '1' if it does
}
\description{
Checking if a given matrix fulfills the Monge property
}
