% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{costMatrix}
\alias{costMatrix}
\title{Cost Matrix}
\usage{
costMatrix(x, y, p = 1, q = 2, wfr = FALSE)
}
\arguments{
\item{x}{A numeric matrix. Each row corresponds to the coordinates of one point in the first point cloud.}

\item{y}{A numeric matrix. Each row corresponds to the coordinates of one point in the second point cloud}

\item{p}{exponent to apply to the distance}

\item{q}{parameter for the minkowski metric. standard p = 2 give the minkowski metric.}

\item{wfr}{computing the wasserstein fisher rao distance}
}
\value{
The distance matrix between the points. The rows correspond to the points in x, the columns to the
points in y
}
\description{
Calculates the cost matrix between points.
}
