% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{costMatrixAlt}
\alias{costMatrixAlt}
\title{Distance matrix}
\usage{
costMatrixAlt(x, y, method = "euclidean")
}
\arguments{
\item{x}{The first coordinates.}

\item{y}{The second coordinates.}

\item{method}{Determines which distance function to use for the computation. The default value is "euclidean" but every
method given for stats::dist can be used.}
}
\value{
The distance matrix.
}
\description{
Calculating the distance matrix for point clouds in 2D.
}
