% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{createCostMatrix}
\alias{createCostMatrix}
\title{Distance matrix}
\usage{
createCostMatrix(x, y, method = "euclidean")
}
\arguments{
\item{x}{The first coordinates.}

\item{y}{The second coordinates.}
}
\value{
The distance matrix.
}
\description{
Calculating the distance matrix for point clouds in 2D.
}
