% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demandExample.R
\docType{data}
\name{demandExample}
\alias{demandExample}
\title{Example demand Distribution}
\format{
A vector containing 1000 numeric values
}
\usage{
demandExample
}
\description{
Example demand Distribution
}
\keyword{datasets}
