% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{findPath}
\alias{findPath}
\title{findPath}
\usage{
findPath(from, to, treeDF)
}
\arguments{
\item{from}{node at which the path starts}

\item{to}{node at which the path ends. Must be in the subtree of 'from'}

\item{treeDF}{tree in data.frame format}
}
\value{
A list of nodes from 'from' to 'to'
}
\description{
Finding the path between a node and a node in its subtree.
}
