% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fVectorDivergence}
\alias{fVectorDivergence}
\title{Computing the divergence functions values}
\usage{
fVectorDivergence(p, u, DivFun, param1, param2 = 0)
}
\arguments{
\item{p}{input vector}

\item{u}{comparision vector}

\item{DivFun}{kind of function to use}

\item{param1}{lambda or alpha}

\item{param2}{beta or 0}
}
\description{
Computing the divergence functions values
}
