% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmMonge.R
\name{genMongeMat}
\alias{genMongeMat}
\title{Monge Matrix Generator

Generates a Monge matrix of given dimensions.}
\usage{
genMongeMat(dimx, dimy)
}
\arguments{
\item{dimx}{number of columns}

\item{dimy}{number of rows}
}
\value{
A Monge matrix of size dimx times dimy.
}
\description{
Monge Matrix Generator

Generates a Monge matrix of given dimensions.
}
