% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmTree.R
\name{getCostMatrix}
\alias{getCostMatrix}
\title{Tree Cost Matrix

Calculating the cost matrix for a given tree.}
\usage{
getCostMatrix(tree)
}
\arguments{
\item{tree}{A tree in list format. Every element in the list is a vector
containing the parent node, the child node and the edge weight.}
}
\value{
The cost matrix for the given tree.
}
\description{
Tree Cost Matrix

Calculating the cost matrix for a given tree.
}
