% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinkhornDivergence.R
\name{grad_legrende}
\alias{grad_legrende}
\title{legendre grad}
\usage{
grad_legrende(lambda, x, DivFun, param1 = 0, param2 = 0)
}
\arguments{
\item{lambda}{num value}

\item{x}{num value}

\item{DivFun}{div fun value}

\item{param1}{num val}

\item{param2}{num val}
}
\description{
legendre grad
}
