% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{gridPlotTransport}
\alias{gridPlotTransport}
\title{A grid plot of the transport plan}
\usage{
gridPlotTransport(transportPlan, import = NULL, export = NULL)
}
\arguments{
\item{transportPlan}{A non negative numeric matrix that indicates where the mass is
transported. The value at point \eqn{\[i,j\]} is the amount of mass transported from
supply point \eqn{i} to demand point \eqn{j}.}

\item{import}{(optional) A non negative numeric vector that give the amount of mass created at each
demand point. It length has to be equal to the number of columns in the transport matrix.}

\item{export}{(optional) A non negative numeric vector that give the amount of mass destroyed  at each
supply point. It length has to be equal to the number of rows in the transport matrix.}
}
\description{
This function creates a grid plot of a transport plan. Import and export vectors can
be given as additional arguments. These will be plotted along the sides of the grid 
plot in order to indicate where mass is added or removed. Hence, the mass shown in
each row or column is equal to the supply or demand vectors.
}
\examples{

transport <- matrix(runif(9), nrow = 3)
import <- runif(3)
export <- runif(3)

gridPlotTransport(transport)
gridPlotTransport(transport, import, export)

}
