% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hausdorff_Vec_Rcpp}
\alias{Hausdorff_Vec_Rcpp}
\title{The symmetric stabilized Scaling Algorithm}
\usage{
Hausdorff_Vec_Rcpp(
  costMatrix,
  distribution,
  f,
  lambda,
  param1,
  param2,
  Div,
  eps
)
}
\arguments{
\item{costMatrix}{A numeric matrix.}

\item{distribution}{num distrie}

\item{f}{A numeric vector}

\item{lambda}{Parameter for the supply proxdiv function}

\item{param1}{numeric Value}

\item{param2}{numeric value}

\item{Div}{Parameter indicating the divergence function to use for the supply proxdiv function}

\item{eps}{A numeric vector of decreasing epsilon values.}
}
\value{
The optimal transport plan
}
\description{
C++ implementation of the log-domain stabilized Version of the Scaling
Algorithm.
}
