% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inital}
\alias{inital}
\title{computing the inital values for f and g}
\usage{
inital(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
the function value
}
\description{
computing the inital values for f and g
}
