% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambertWFunctionLog}
\alias{lambertWFunctionLog}
\title{computing the lambert w function}
\usage{
lambertWFunctionLog(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
the function value
}
\description{
computing the lambert w function
}
