% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LSE}
\alias{LSE}
\title{The Log-Sum-Exp reduction of the vector u}
\usage{
LSE(u)
}
\arguments{
\item{u}{a numeric vector}
}
\value{
log(sum_{i=1, N}(exp(u_i)))
}
\description{
The Log-Sum-Exp reduction of the vector u
}
