% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Monge_Rcpp}
\alias{Monge_Rcpp}
\title{C++ implementation of the Monge algorithm}
\usage{
Monge_Rcpp(costMatrix, supply, demand, constructionCost, destructionCost)
}
\arguments{
\item{costMatrix}{A numeric matrix that fulfills the Monge property}

\item{supply}{A numeric vector. Its length must be equal to the numbers of rows in the cost matrix.}

\item{demand}{A numeric vector. Its length must be equal to the number of column in the cost matrix.}

\item{constructionCost}{The cost of creating mass at any point.}

\item{destructionCost}{The cost of destructing mass at any point.}
}
\value{
It returns a list of the transport cost, three numeric vectors that indicate the optimal transport path, and two vectors that indicate
    where mass is created and where it is destructed.
}
\description{
This function calculates the optimal transport cost and transport
plan for unbalanced optimal transport problems with a Monge cost matrix
}
