% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmMonge.R
\name{mongeAlgorithm}
\alias{mongeAlgorithm}
\title{Monge Unbalanced Optimal Transport Algorithm}
\usage{
mongeAlgorithm(costMatrix, supply, demand, constructionCost, destructionCost)
}
\arguments{
\item{costMatrix}{A numeric Monge matrix.}

\item{supply}{A numeric supply vector}

\item{demand}{A numeric demand vector}

\item{constructionCost}{A numeric construction cost value}

\item{destructionCost}{A numeric destruction cost value}
}
\value{
A list containing the optimal transport cost, plan, import vector
and export vector
}
\description{
This calculates the optimal transport cost and optimal transport path of an
unbalanced optimal transport problem. The cost matrix has to fulfill the Monge
property.
}
