% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmMonge.R
\name{mongeTransport}
\alias{mongeTransport}
\title{Monge Unbalanced Optimal Transport Algorithm}
\usage{
mongeTransport(costMatrix, supply, demand, constructionCost, destructionCost)
}
\arguments{
\item{costMatrix}{A numeric matrix fulfilling the Monge property}

\item{supply}{A numeric supply vector}

\item{demand}{A numeric demand vector}

\item{constructionCost}{A numeric construction cost value}

\item{destructionCost}{A numeric destruction cost value}
}
\value{
A list containing the optimal transport cost, plan, import vector
and export vector
}
\description{
This calculates the optimal transport cost and optimal transport path of an
unbalanced optimal transport problem. The cost matrix has to fulfill the Monge
property.
}
\examples{
# generating random data
mongeMatrix <- genMongeMat(30,30)

supply <- runif(30,0,1)
demand <- runif(30,0,1)

createCost <- runif(1,0,0.2)
destructCost <- runif(1,0,0.3)

# calculating the optimal transport cost and plan
res <- mongeTransport(mongeMatrix, supply, demand, createCost, destructCost)

# plotting the transport plan
plotUOTP(res$transportPlan, res$import, res$export)

}
