% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{nextLayer}
\alias{nextLayer}
\title{Next Layer}
\usage{
nextLayer(treeDF, coordinates, node, layer)
}
\arguments{
\item{treeDF}{A tree in data.frame format}

\item{coordinates}{The coordinates of the tree nodes as data.frame}

\item{node}{The current node index}

\item{layer}{The current layer}
}
\value{
data.frame with coordinates and information for all tree nodes
}
\description{
This calculates the coordinates of the tree nodes in the next layer.
}
