% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Nw_Corner_Rule}
\alias{Nw_Corner_Rule}
\title{C++ implantation of the North-West Corner Rule}
\usage{
Nw_Corner_Rule(costMatrix, supply, demand, iList, jList, weightList)
}
\arguments{
\item{costMatrix}{A numeric matrix.}

\item{supply}{A numeric supply vector.}

\item{demand}{A numeric demand vector.}

\item{iList}{A numeric vector holding the row index of the transport path}

\item{jList}{A numeric vector holding the row index of the transport path}

\item{weightList}{A numeric vector holding the transport path weights}
}
\value{
The transport cost.
}
\description{
This function calculates the optimal transport cost for a balanced optimal
transport problem if its cost matrix fulfills the Monge property. Otherwise,
it calculates a feasible solution for the transport problme.
}
