% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallelVectorLse}
\alias{parallelVectorLse}
\title{The Sinkhorn Algorithm}
\usage{
parallelVectorLse(logSup, f, costMatrix, eps, Ny)
}
\arguments{
\item{logSup}{A numeric vector}

\item{f}{f}

\item{costMatrix}{cm}

\item{eps}{eps}

\item{Ny}{y}
}
\description{
C++ implementation of the Sinkhorn Algorithm.
}
