% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmMonge.R
\name{pathToPlan}
\alias{pathToPlan}
\title{pathToPlan}
\usage{
pathToPlan(iList, jList, wList, dimX, dimY)
}
\arguments{
\item{iList}{A numeric vector of i indexes}

\item{jList}{A numeric vector of j indexes}

\item{wList}{A numeric vector of weights}

\item{dimX}{The number of rows in the transport matrix}

\item{dimY}{The number of columns in the transport matrix}
}
\value{
A dimX times dimY transport matrix
}
\description{
Computes the transport plan from a given transport path.
}
