% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot1DTransport}
\alias{plot1DTransport}
\title{Plotting dense 1D transport}
\usage{
plot1DTransport(transportPlan, supplyList, demandList)
}
\arguments{
\item{transportPlan}{A non negative numeric matrix giving the mass transport. The value at \eqn{(i,j)}
gives the mass transported from supply point \eqn{i} to demand point \eqn{j}.}

\item{supplyList}{A list containing the non negative supply measure and the underlying discretization as vectors.}

\item{demandList}{A list containing the non negative demand measure and the underlying discretization as vectors.}
}
\description{
A function to plot optimal transport between discretizations of continuous supply and demand measures in 1D.
}
