% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot2DTransport}
\alias{plot2DTransport}
\title{Plotting dense 1D transport}
\usage{
plot2DTransport(transportMap, supplyList, demandList)
}
\arguments{
\item{transportMap}{A numeric matrix.}

\item{supplyList}{A supply list containing the divergence to use (either "KL" or "TV"),
a numeric supply vector, the reference measure as numeric vector and the
value for the lambda parameter.}

\item{demandList}{A demand list containing the divergence to use (either "KL" or "TV"),
a numeric demand vector, the reference measure as numeric vector and the
value for the lambda parameter.}
}
\description{
Plotting dense 1D transport
}
