% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotTransportByCost}
\alias{plotTransportByCost}
\title{Plotting transport between points given by a distance matrix}
\usage{
plotTransportByCost(
  distanceMatrix,
  transportPlan = NULL,
  supply = NULL,
  demand = NULL,
  creationDestructionCost = rep(0, length(x))
)
}
\arguments{
\item{distanceMatrix}{A numeric distance matrix.}

\item{transportPlan}{A numeric matrix.}

\item{supply}{A numeric vector.}

\item{demand}{A numeric vector.}

\item{creationDestructionCost}{A numeric vector giving the cost for mass creation and destruction.}
}
\description{
Plotting transport between points given by a distance matrix
}
