% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotTree}
\alias{plotTree}
\title{Tree plot}
\usage{
plotTree(tree, tList = NULL, supply = NULL, demand = NULL)
}
\arguments{
\item{tree}{A tree in list format: The root node followed by multiple vectors of the form
(parent, child, weight)}

\item{tList}{The transport list (optional)}

\item{supply}{A numeric supply vector (optional)}

\item{demand}{A numeric demand vector (optional)}
}
\description{
Plotting a tree structure. Supply and demand values as well as the transport plan
can be plotted in one graph.
}
\examples{
tree <- list(1, c(1, 2, 1), c(2, 3, 1), c(3, 4, 1), c(3, 5, 1),
c(2, 6, 1), c(1, 7, 1), c(7, 8, 1), c(7, 9, 1),
c(9, 10, 1), c(9, 11, 1), c(11, 12, 1), c(11, 13, 1))

plotTree(tree)


supply <- c(0,0,1,2,0,0,0,0,0,0,0,0,0)
demand <- c(0,0,0,0,0,1,0,1,0,0,0,1,1)

plotTree(tree, supply = supply, demand = demand)


tList = list(c(3,6,1), c(4,8,1))
plotTree(tree, tList, supply, demand)


}
