% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotUOTP}
\alias{plotUOTP}
\title{Plotting an unbalanced optimal transport plan}
\usage{
plotUOTP(transportPlan, import = NULL, export = NULL)
}
\arguments{
\item{transportPlan}{A numeric matrix.}

\item{import}{A numeric vector.}

\item{export}{A numeric vector.}
}
\description{
Plotting an unbalanced optimal transport plan
}
