% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{positionsByEigenvalue}
\alias{positionsByEigenvalue}
\title{Coordinates by distance matrix}
\usage{
positionsByEigenvalue(c)
}
\arguments{
\item{c}{A numeric distance matrix.}
}
\value{
The coordinates based on the distance matrix.
}
\description{
Calculating the coordinates of the supply and demand points using
the eigenvalue decomposition of the distance matrix.
}
