% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proxdiv}
\alias{proxdiv}
\title{The proxdiv operator}
\usage{
proxdiv(lambda, p, s, u, eps, DivFun)
}
\arguments{
\item{lambda}{Regularization parameter}

\item{p}{A numeric vector}

\item{s}{A numeric vector}

\item{u}{A numeric vector}

\item{eps}{The epsilon value}

\item{DivFun}{A numeric value indicating the function to be used.'1' gives
the proxdiv operator for the Kullback-Leibler divergence and '2' the opterator
for the total variation.}
}
\value{
A vector holding the proxdiv evaluation
}
\description{
The proxdiv operators for different divergences in the form of 'lambda * DivFun(.|p)'
Implemented are the operators for the Kullback-Leibler divergence and total variation.
}
