% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{quadCost}
\alias{quadCost}
\title{Quadratic Cost Matrix}
\usage{
quadCost(X, Y)
}
\arguments{
\item{X}{A numeric vector.}

\item{Y}{A numeric vector.}
}
\value{
The quadratic cost matrix.
}
\description{
Calculating the quadratic cost matrix.
}
