% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinkhornDivergence.R
\name{regularizedf_ot}
\alias{regularizedf_ot}
\title{regularized output}
\usage{
regularizedf_ot(
  supplyList,
  demandList,
  eps,
  iterMax = 100,
  tol = 0.001,
  method = "euclidean",
  exp = 1,
  p = 2,
  wfr = FALSE,
  Cxy = NULL
)
}
\arguments{
\item{supplyList}{sup list}

\item{demandList}{dem list}

\item{eps}{num val}

\item{iterMax}{num val}

\item{tol}{num val}

\item{method}{cost method}

\item{exp}{cost exponent}

\item{p}{cost parameter}

\item{wfr}{wfr param}

\item{Cxy}{cost matrix}
}
\description{
regularized output
}
