% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedTransport.R
\name{regularizedTransport}
\alias{regularizedTransport}
\title{Regularized Unbalanced Optimal Transport}
\usage{
regularizedTransport(
  supplyList,
  demandList,
  supplyDivList = list("KL", 0.5),
  demandDivList = list("KL", 0.5),
  epsVector = c(0.5),
  maxIteration = 5000,
  tol = 1e-08,
  p = 1,
  q = 2,
  wfr = FALSE,
  costMatrix = NULL,
  duals = FALSE,
  algorithm = "scaling"
)
}
\arguments{
\item{supplyList}{a}

\item{demandList}{a}

\item{supplyDivList}{a}

\item{demandDivList}{a}

\item{epsVector}{a}

\item{maxIteration}{a}

\item{tol}{a}

\item{p}{a}

\item{q}{a}

\item{wfr}{a}

\item{costMatrix}{a}

\item{duals}{a}

\item{algorithm}{a}
}
\description{
Regularized Unbalanced Optimal Transport
}
