% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sinkhorn_Rcpp}
\alias{Sinkhorn_Rcpp}
\title{The stabilized Scaling Algorithm}
\usage{
Sinkhorn_Rcpp(
  costMatrix,
  supply,
  demand,
  lambdaSupply,
  param1Supply,
  param2Supply,
  lambdaDemand,
  param1Demand,
  param2Demand,
  DivSupply,
  DivDemand,
  iterMax,
  eps,
  tol
)
}
\arguments{
\item{costMatrix}{A numeric matrix.}

\item{supply}{A numeric vector}

\item{demand}{A numeric vector}

\item{lambdaSupply}{Parameter for the supply proxdiv function}

\item{param1Supply}{numeric Value}

\item{param2Supply}{numeric value}

\item{lambdaDemand}{Parameter for the demand proxdiv function}

\item{param1Demand}{numeric Value}

\item{param2Demand}{numeric value}

\item{DivSupply}{Parameter indicating the divergence function to use for the supply proxdiv function}

\item{DivDemand}{Parameter indicating the divergence function to use for the demand proxdiv function}

\item{iterMax}{Maximum number of iterations}

\item{eps}{A numeric vector of decreasing epsilon values.}

\item{tol}{numeric value}
}
\value{
The optimal transport plan
}
\description{
C++ implementation of the log-domain stabilized Version of the Scaling
Algorithm.
}
