% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmSinkhorn.R
\name{sinkhornAlgorithmFromCost}
\alias{sinkhornAlgorithmFromCost}
\title{The Sinkhorn Algorihm for unbalanced optimal transport}
\usage{
sinkhornAlgorithmFromCost(
  costMatrix,
  supplyList,
  demandList,
  maxIteration,
  eps
)
}
\arguments{
\item{costMatrix}{A numeric matrix.}

\item{supplyList}{A supply list containing the divergence to use (either "KL" or "TV"),
a numeric supply vector, the reference measure as numeric vector and the
value for the lambda parameter.}

\item{demandList}{A demand list containing the divergence to use (either "KL" or "TV"),
a numeric demand vector, the reference measure as numeric vector and the
value for the lambda parameter.}

\item{maxIteration}{A numeric value for the maximum number of iterations.
The default value is 20000.}

\item{eps}{vector of epsilon values to use}
}
\description{
The Sinkhorn Algorihm for unbalanced optimal transport
}
