% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplyExample.R
\docType{data}
\name{supplyExample}
\alias{supplyExample}
\title{Example supply Distribution}
\format{
A vector containing 1000 numeric values
}
\usage{
supplyExample
}
\description{
Example supply Distribution
}
\keyword{datasets}
