% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{treegkr_Rcpp}
\alias{treegkr_Rcpp}
\title{The tree metric unbalanced optimal transport algorithm}
\usage{
treegkr_Rcpp(tree, supply, demand, creation, destruction)
}
\arguments{
\item{tree}{A tree structure given in list form. Each entry in the list represents
an edge: (first node, second node, edge weight)}

\item{supply}{The supply vector.}

\item{demand}{The demand vector.}

\item{creation}{A numeric vector giving the creation cost at each node.}

\item{destruction}{A numeric vector giving the destruction cost at each node.}
}
\value{
A list containing the optimal transport cost and the import vector.
}
\description{
This function makes the unbalanced optimal transport algorithm for tree metrics
accessible from R. It calculated the optimal transport cost and the import vector.
}
