% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalancedTransport.R
\docType{package}
\name{unbalancedTransport}
\alias{unbalancedTransport}
\title{UnbalancedTransport: A package for solving unbalanced optimal transport problems}
\description{
The package provides algorithms for solving different types of unbalanced OT problems.
}
\section{Regularized unbalanced optimal transport}{

Text
}

\section{Unbalanced optimal transport on trees}{

Text
}

\section{Linear unbalanced optimal transport}{

Text
}

