% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorDivergence}
\alias{vectorDivergence}
\title{Computing the divergence functions values}
\usage{
vectorDivergence(r, s, DivFun, param1, param2 = 0)
}
\arguments{
\item{r}{input vector}

\item{s}{comparision vector}

\item{DivFun}{kind of function to use}

\item{param1}{lambda or alpha}

\item{param2}{beta or 0}
}
\description{
Computing the divergence functions values
}
