% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{wfrCost}
\alias{wfrCost}
\title{Wasserstein-Fisher-Rao Cost Matrix}
\usage{
wfrCost(X, Y)
}
\arguments{
\item{X}{A numeric vector.}

\item{Y}{A numeric vector.}
}
\value{
The WFR cost matrix.
}
\description{
Calculating the Wasserstein-Fisher-Rao cost matrix for the absolute value
distance function between two 1D sets of points.
}
