% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalancedExtensionSolver.R
\name{BalancedExtensionSolver}
\alias{BalancedExtensionSolver}
\title{Extendes the UOT Problemt to the balanced case and solves it using the known methods}
\usage{
BalancedExtensionSolver(
  supplyList,
  demandList,
  costMethod,
  costExp,
  costP,
  wfr = FALSE
)
}
\arguments{
\item{supplyList}{list of supply values}

\item{demandList}{list of demand values}

\item{costMethod}{method to use for the cost matrix}

\item{costExp}{exponent for the cost matrix}

\item{costP}{parameter to use for the minkowsi cost matrix}

\item{wfr}{using wfr cost matrix??}
}
\description{
Extendes the UOT Problemt to the balanced case and solves it using the known methods
}
