% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{costMatrix}
\alias{costMatrix}
\title{Cost Matrix}
\usage{
costMatrix(x, y, method = "euclidean", p = 1)
}
\arguments{
\item{x}{A numeric matrix. Each row corresponds to the coordinates of one point in the first point cloud.}

\item{y}{A numeric matrix. Each row corresponds to the coordinates of one point in the second point cloud}

\item{method}{Determines which distance function to use for the computation.  Currently implemented are:
1. Euclidean distance:
2. Wassserstein Fisher Rao distance:}
}
\value{
The distance matrix between the points. The rows correspond to the points in x, the columns to the
points in y
}
\description{
Calculates the cost matrix between points.
}
