% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmTree.R, R/treeAlgorithm.R
\name{getCostMatrix}
\alias{getCostMatrix}
\title{Tree Cost Matrix

Calculating the cost matrix for a given tree.}
\usage{
getCostMatrix(tree)

getCostMatrix(tree)
}
\arguments{
\item{tree}{a tree in list format}
}
\value{
The cost matrix for the given tree.

the cost matrix for the given tree
}
\description{
Tree Cost Matrix

Calculating the cost matrix for a given tree.

getCostMatrix
}
