% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinkhornDivergence.R
\name{hausdorff_divergence}
\alias{hausdorff_divergence}
\title{Hausdorff Divergence}
\usage{
hausdorff_divergence(
  supplyList,
  demandList,
  supplyDivList,
  demandDivList,
  epsVector,
  iterMax = 1000,
  tol = 0.001,
  exp = 1,
  p = 2,
  wfr = FALSE,
  Cxy = NULL,
  Cxx = NULL,
  Cyy = NULL
)
}
\arguments{
\item{supplyList}{A list containing the supply measure and if the cost matrix is not provided the support of the supply distribution.}

\item{demandList}{A list containing the demand measure and if the cost matrix is not provided the support of the demand distribution.}

\item{supplyDivList}{A list containing the information about the divergence function for the supply measure. The first element is the
abbreviation of the divergence function. The following values are available: 

\itemize{
  \item "KL" for the Kullback-Leibner divergence \eqn{F = \lambda \cdot KL()}{F1 = lambda * KL()}
  \item "TV" for total variation divergence \eqn{F = \lambda \cdot TV()}{F1 = lambda * TV()}
  \item "RG" for the range constraint 
  \item "Power" for the Power divergence with conjugate exponent r: \eqn{F_1 = \lambda \cdot Power_{r}()}{F1 = lambda * Power_r()}
  \item "Hellinger" for the Power divergence with conjugate exponent r = -1: \eqn{F_1 = \lambda \cdot Power_{-1}()}{F1 = lambda * Power_-1()}
  \item "Berg" for the Power divergence with conjugate exponent r = 0: \eqn{F_1 = \lambda \cdot Power_0()}{F1 = lambda * Power_0()}
}

The next elements in supplyDivList are the additional parameters for the divergence functions. For "KL", "TV", "Hellinger"
and "Berg" the only required parameter is the regularization parameter \eqn{\lambda}{lambda}. The "RG" divergence function requires two parameters
alpha and beta with \eqn{0 \leq \alpha \leq \beta}{0 <= alpha <= beta} that define the lower and upper bounds. Lastly, "Power" requires
two parameters as well. The first is the regularization parameter \eqn{\lambda}{lambda}, followed by the conjugate exponent \eqn{r}.}

\item{demandDivList}{A list containing the information about the divergence function for the demand measure in the same form as the supplyDivList.}

\item{epsVector}{A numeric value or vector holding the regularization parameter. If a vector is given, epsilon scaling will be performed.}

\item{iterMax}{(optional) The maximum number of algorithm iterations. The default value is 5000}

\item{tol}{(optional) A numeric value. If the change of the dual variables from one step to the next is smaller than this value. The algorithm
is terminated as it is already converged. The default value is 1e-5.}

\item{exp}{(optional) The exponent that is applied to the cost matrix. Can be used to compute quadratic cost. The default value is 1.}

\item{p}{(optional) Parameter for the L_p cost function. Can be omitted if either "euclidean" or "maximum" is used. The default value is 2.}

\item{wfr}{(optional) Set to "TRUE" to calculate the cost matrix for the Wasserstein-Fisher-Rao distance
\eqn{c(x,y) = -\log(\cos^2_+(d(x,y)))}{c(x,y) = -log(cos_+(d(x,y)²))}. Default value is "FALSE".}

\item{Cxy}{(optional) A cost matrix for transport between the supply and demand distributions.}

\item{Cxx}{(optional) A cost matrix for transport between the supply and supply distributions.}

\item{Cyy}{(optional) A cost matrix for transport between the demand and demand distributions.}
}
\value{
The Hausdorff divergence.
}
\description{
Calculating the Hausdorff divergence for regularized unbalanced optimal transport.
}
\details{
This function calculates the Hausdorff divergence using the dual optimal optimal potentials
calculated by the \code{\link[unbalancedTransport]{regularizedTransport}} function.
The Hausdorff divergence is defined as 
\eqn{H_\varepsilon(\alpha, \beta) = <\alpha - \beta, \nabla F_\varepsilon(\alpha) - 
\nabla F_\varepsilon(\beta)>}{H_eps(a,b) = <a-b, nabla(F_eps(a)) - nabla(F_eps(b)) >}  
with supply and demand distributions \eqn{\alpha}{a} and \eqn{\beta}{b} and
\eqn{F_\varepsilon(\alpha) = -\frac{1}{2} OT_\varepsilon(\alpha,\alpha) + 
\frac{\varepsilon}{2}(m(\alpha)^2)}{F_eps(a) = -0.5 OT_eps(a,a) + 0.5eps * m(a)^2}.
}
\examples{

\donttest{
I <- 1000
J <- 1000
X <- seq(0,1,length.out = I)
Y <- seq(0,1,length.out = J)
p <- supplyExample
q <- demandExample

supply <- list(p,X)
demand <- list(q,Y)

eps <- 1e-3 
supplyDiv <- list("KL", 1.05)
demandDiv <- list("KL", 1.05)

hausdorff_divergence(supply, demand, supplyDiv, demandDiv, eps, exp = 2)
}



}
\references{
\insertRef{Sejourne2019}{unbalancedTransport}
}
