% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_vectors}
\alias{init_vectors}
\title{Inital values}
\usage{
init_vectors(
  lambda,
  costMatrix,
  distribution,
  secDistribution,
  DivFun,
  param1,
  param2,
  Nx,
  Ny,
  eps
)
}
\arguments{
\item{lambda}{Regularization parameter}

\item{DivFun}{A numeric value indicating the function to be used.'1' gives
the proxdiv operator for the Kullback-Leibler divergence and '2' the opterator
for the total variation.}

\item{param1}{num value}

\item{param2}{num value}

\item{eps}{The epsilon value}

\item{p}{A numeric vector}
}
\value{
A vector holding the proxdiv evaluation
}
\description{
The aprix operators for different divergences in the form of 'lambda * DivFun(.|p)'
Implemented are the operators for the Kullback-Leibler divergence and total variation.
}
