% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{inseg}
\alias{inseg}
\title{Interval check}
\usage{
inseg(x, s1, s2)
}
\arguments{
\item{x}{A numeric value.}

\item{s1}{The left boundary of a 1D interval.}

\item{s2}{The right boundary of a 1D interval.}
}
\value{
1 if x is in [s1,s2] and 0 if it is not.
}
\description{
This checks if a value is in a given interval
}
