% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotTransportByCost}
\alias{plotTransportByCost}
\title{Plotting transport between single points.}
\usage{
plotTransportByCost(
  transportPlan,
  supply = NULL,
  demand = NULL,
  distanceMatrix = NULL,
  creationDestructionCost = rep(0, length(x))
)
}
\arguments{
\item{transportPlan}{A non negative numeric matrix that indicates where the mass is
transported. The value at point \eqn{\[i,j\]} is the amount of mass transported from
supply point \eqn{i} to demand point \eqn{j}.}

\item{supply}{(optional) A non negative numeric vector giving the mass supply at each point. The value at 
position \eqn{i} give the supply at point \eqn{i}.}

\item{demand}{(optional) A non negative numeric vector giving the mass demand at each point. The value at 
position \eqn{i} give the supply at point \eqn{i}.}

\item{distanceMatrix}{(optional) A non negative matrix containing the distances between all points. It is used
to calcuated the positions of the points if they are not explicitly given.}

\item{creationDestructionCost}{(optional) A numeric vector giving the cost for mass creation and destruction.}
}
\description{
Visualizing mass transport between point clouds. The coordinates can either be given directly or calculated from a given
distance matrix.
}
