% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotTree}
\alias{plotTree}
\title{Plotting transport on trees}
\usage{
plotTree(tree, tList = NULL, supply = NULL, demand = NULL)
}
\arguments{
\item{tree}{A tree structure in list format: The first element is the index of the root node
followed by multiple vectors defining the edges of the tree. Each of these vectors has to be 
of the form \eqn{(parent, child, weight)}.}

\item{tList}{(optional) The mass transport as list. Each element is a vector of the form
\eqn{(source, target, mass)}.}

\item{supply}{(optional) A non negative numeric vector giving the mass supply at each tree node. The value at 
position \eqn{i} give the supply at node \eqn{i}.}

\item{demand}{(optional) A non negative numeric vector giving the mass demand at each tree node. The value at 
position \eqn{i} give the supply at node \eqn{i}.}
}
\description{
This function visualizes the transport of mass on a tree.
}
\examples{
tree <- list(1, c(1, 2, 1), c(2, 3, 1), c(3, 4, 1), c(3, 5, 1),
c(2, 6, 1), c(1, 7, 1), c(7, 8, 1), c(7, 9, 1),
c(9, 10, 1), c(9, 11, 1), c(11, 12, 1), c(11, 13, 1))

plotTree(tree)


supply <- c(0,0,1,2,0,0,0,0,0,0,0,0,0)
demand <- c(0,0,0,0,0,1,0,1,0,0,0,1,1)

plotTree(tree, supply = supply, demand = demand)


tList = list(c(3,6,1), c(4,8,1))
plotTree(tree, tList, supply, demand)

}
