% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinkhornDivergence.R
\name{regularized_ot}
\alias{regularized_ot}
\title{Regularized transport cost

A function to calculate the regularized transport cost using the dual optimal optimal potentials.

The regularized transport cost is defined as \eqn{OT_\varepsilon = sup_{f,g} -<\alpha, \phi^*(-f)> - 
<\beta, \phi^*(-g)> - \varepilon <\alpha \times \beta , \exp{(f+g-C)/\varepsilon}- 1>}

with supply and demand distributions \eqn{\alpha}{a} and \eqn{\beta}{b}, optimal dual potentials \eqn{f} and \eqn{g}, 
Legendre transforms of the divergence function \eqn{\phi^*}, cost matrix \eqn{C} and regularization parameter \eqn{\varepsilon}{eps}.

The function uses the Sinkhorn algorithm to compute the optimal dual potentials.



[REF]}
\usage{
regularized_ot(
  supplyList,
  demandList,
  eps,
  maxIteration = 100,
  tol = 0.001,
  algorithm = "sinkhorn",
  method = "euclidean",
  exp = 1,
  p = 2,
  wfr = FALSE,
  Cxy = NULL
)
}
\arguments{
\item{supplyList}{A list containing the information about the supply measure, 
divergence function and parameter. The first element is the supply measure itself,
the second the abbreviation of the divergence function ("KL" for Kullback Leibner, 
"TV" for total variation and "RG" for range constraint) followed by the parameters 
needed for the divergence function: 

In case of the KL divergence \eqn{F_1 = \lambda \cdot KL()}{F1 = lambda * KL()} and the 
regularization parameter \eqn{\lambda}{lambda} has to be provided.

The same structure is used for the TV divergence: \eqn{F_1 = \lambda \cdot TV()}{F1 = lambda * TV()} and the 
regularization parameter \eqn{\lambda}{lambda} has to be provided.

The divergence function associated with the range constraint needs two parameters, that define the upper and 
lower bound. 

If the cost matrix is not provided, the support of the supply measure has to be provied as last element in the list. This can
be omitted if a cost matrix is passed as argument.}

\item{demandList}{A list containing the information about the demand measure. It has to have the same structure as the supplyList.}

\item{eps}{The regularization parameter.}

\item{maxIteration}{The maximum number of iterations.}

\item{tol}{(optional) A numeric value. If the change of the dual variables from one step to the next is smaller than this value. The algorithm
is terminated as it is already converged. The default value is 1e-10.}

\item{algorithm}{(optional) Determines which algorithm to use. Possible values are "sinkhorn"
for \code{\link[unbalancedTransport]{sinkhornAlgorithm}} and "scaling" for \code{\link[unbalancedTransport]{scalingAlgorithm}}}

\item{method}{(optional) Determines the method that is used to compute the cost matrix.
\itemize{
\item "euclidean"
\item "minkowski"
\item "maximum" 
}
The default value is "euclidean".}

\item{exp}{(optional) The exponent that is applied to the cost matrix. Can be used to compute quadratic cost. The default value is 1.}

\item{p}{(optional) Parameter for the minkwski cost function. Can be omitted if either "euclidean" or "maximum" is used. The default value is 2.}

\item{wfr}{(optional) Computes the cost matrix needed for the Wasserstein-Fisher-Rao distance \eqn{c(x,y) = -\log(\cos^2_+(d(x,y)))}{c(x,y) = -log(cos_+(d(x,y)²))}.
The default value is "false".}

\item{Cxy}{(optional) Instead of having the algorithm compute the cost matrix, a custom cost matrix can be passed to the algorithm.}
}
\description{
Regularized transport cost

A function to calculate the regularized transport cost using the dual optimal optimal potentials.

The regularized transport cost is defined as \eqn{OT_\varepsilon = sup_{f,g} -<\alpha, \phi^*(-f)> - 
<\beta, \phi^*(-g)> - \varepilon <\alpha \times \beta , \exp{(f+g-C)/\varepsilon}- 1>}

with supply and demand distributions \eqn{\alpha}{a} and \eqn{\beta}{b}, optimal dual potentials \eqn{f} and \eqn{g}, 
Legendre transforms of the divergence function \eqn{\phi^*}, cost matrix \eqn{C} and regularization parameter \eqn{\varepsilon}{eps}.

The function uses the Sinkhorn algorithm to compute the optimal dual potentials.



[REF]
}
