% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmScaling.R
\name{scalingAlgorithm}
\alias{scalingAlgorithm}
\title{The log-domain stabilized Scaling Algorithm}
\usage{
scalingAlgorithm(
  supplyList,
  demandList,
  maxIteration,
  epsVector,
  method = "euclidean",
  exp = 1,
  p = 2,
  wfr = FALSE
)
}
\arguments{
\item{supplyList}{A supply list containing the divergence to use (either "KL" or "TV"),
a numeric supply vector, the reference measure as numeric vector and the
value for the lambda parameter.}

\item{demandList}{A demand list containing the divergence to use (either "KL" or "TV"),
a numeric demand vector, the reference measure as numeric vector and the
value for the lambda parameter.}

\item{maxIteration}{A numeric value for the maximum number of iterations.
The default value is 20000.}

\item{epsVector}{vector of epsilon values to use}

\item{method}{distance method}

\item{exp}{exponent for distance}

\item{p}{p norm value}

\item{wfr}{wfr distance?}
}
\description{
The log-domain stabilized Scaling Algorithm
}
