% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmScaling.R
\name{scalingAlgorithmFromCost}
\alias{scalingAlgorithmFromCost}
\title{The log-domain stabilized Scaling Algorithm}
\usage{
scalingAlgorithmFromCost(
  costMatrix,
  supplyList,
  demandList,
  maxIteration,
  epsVector,
  tol = 1e-10
)
}
\arguments{
\item{costMatrix}{A numeric matrix.}

\item{supplyList}{A supply list containing the divergence to use (either "KL" or "TV"),
a numeric supply vector, the reference measure as numeric vector and the
value for the lambda parameter.}

\item{demandList}{A demand list containing the divergence to use (either "KL" or "TV"),
a numeric demand vector, the reference measure as numeric vector and the
value for the lambda parameter.}

\item{maxIteration}{A numeric value for the maximum number of iterations.
The default value is 20000.}

\item{epsVector}{vector of epsilon values to use}

\item{tol}{num value}
}
\description{
The log-domain stabilized Scaling Algorithm
}
