% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinkhornDivergence.R
\name{sinkhorn_divergence}
\alias{sinkhorn_divergence}
\title{Sink div}
\usage{
sinkhorn_divergence(
  supplyList,
  demandList,
  eps,
  iterMax = 1000,
  tol = 0.001,
  method = "euclidean",
  exp = 1,
  p = 2,
  wfr = FALSE,
  Cxy = NULL,
  Cxx = NULL,
  Cyy = NULL
)
}
\arguments{
\item{supplyList}{sup list}

\item{demandList}{dem list}

\item{eps}{num val}

\item{iterMax}{num val}

\item{tol}{num val}

\item{method}{cost method}

\item{exp}{cost exponent}

\item{p}{cost parameter}

\item{wfr}{wfr param}

\item{Cxy}{cost matrix}

\item{Cxx}{cost matrix}

\item{Cyy}{cost matrix}
}
\description{
Sink div
}
