% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinkhornDivergence.R
\name{sinkhorn_divergence}
\alias{sinkhorn_divergence}
\title{Sinkhorn Divergence}
\usage{
sinkhorn_divergence(
  supplyList,
  demandList,
  supplyDivList,
  demandDivList,
  epsVector,
  iterMax = 10000,
  tol = 1e-05,
  p = 1,
  q = 2,
  wfr = FALSE,
  Cxy = NULL,
  Cxx = NULL,
  Cyy = NULL
)
}
\arguments{
\item{supplyList}{A list containing the supply measure and if the cost matrix is not provided the support of the supply distribution.}

\item{demandList}{A list containing the demand measure and if the cost matrix is not provided the support of the demand distribution.}

\item{supplyDivList}{A list containing the information about the divergence function for the supply measure. The first element is the
abbreviation of the divergence function. The following values are available: 

\itemize{
  \item "KL" for the Kullback-Leibner divergence \eqn{F = \lambda \cdot KL()}{F1 = lambda * KL()}
  \item "TV" for total variation divergence \eqn{F = \lambda \cdot TV()}{F1 = lambda * TV()}
  \item "RG" for the range constraint 
  \item "Power" for the Power divergence with conjugate exponent r: \eqn{F_1 = \lambda \cdot Power_{r}()}{F1 = lambda * Power_r()}
  \item "Hellinger" for the Power divergence with conjugate exponent r = -1: \eqn{F_1 = \lambda \cdot Power_{-1}()}{F1 = lambda * Power_-1()}
  \item "Berg" for the Power divergence with conjugate exponent r = 0: \eqn{F_1 = \lambda \cdot Power_0()}{F1 = lambda * Power_0()}
}


The next elements in supplyDivList are the additional parameters for the divergence functions. For "KL", "TV", "Hellinger"
and "Berg" the only required parameter is the regularization parameter \eqn{\lambda}{lambda}. The "RG" divergence function requires two parameters
alpha and beta with \eqn{0 \leq \alpha \leq \beta}{0 <= alpha <= beta} that define the lower and upper bounds. Lastly, "Power" requires
two parameters as well. The first is the regularization parameter \eqn{\lambda}{lambda}, followed by the conjugate exponent \eqn{r}}

\item{demandDivList}{A list containing the information about the divergence function for the
demand measure in the same form as the supplyDivList.}

\item{epsVector}{A numeric value or vector holding the regularization parameter. If a vector is given, epsilon scaling will be performed.}

\item{iterMax}{(optional) The maximum number of algorithm iterations. The default value is 10000.}

\item{tol}{(optional) A numeric value. If the change of the dual variables from one step to the next is smaller than this value. The algorithm
is terminated as it is already converged. The default value is 1e-5.}

\item{p}{(optional) The exponent that is applied to the cost matrix. Can be used to compute quadratic cost. The default value is 1.}

\item{q}{(optional) Parameter for the minkwski cost function. Can be omitted if either "euclidean" or "maximum" is used. The default value is 2.}

\item{wfr}{(optional) Set to "TRUE" to calculate the cost matrix for the Wasserstein-Fisher-Rao distance
\eqn{c(x,y) = -\log(\cos^2_+(d(x,y)))}{c(x,y) = -log(cos_+(d(x,y)²))}. Default value is "FALSE".}

\item{Cxy}{(optional) A cost matrix for transport between the supply and demand distributions.}

\item{Cxx}{(optional) A cost matrix for transport between the supply and supply distributions.}

\item{Cyy}{(optional) A cost matrix for transport between the demand and demand distributions.}
}
\value{
The Sinkhorn divergence.
}
\description{
Calculating the Sinkhorn divergence for regularized unbalanced optimal transport.
}
\details{
A function to calculating the Sinkhorn divergence for regularized unbalanced optimal transport.
For supply and demand measures \eqn{\alpha}{a} and \eqn{\beta}{b} the Sinkhorn divergence is defined as 
\eqn{S_{\varepsilon} = OT_\varepsilon(\alpha,\beta) - \frac{1}{2}OT_\varepsilon(\alpha,\alpha) -
\frac{1}{2}OT_\varepsilon(\beta,\beta) + \frac{\varepsilon}{2}(m(\alpha) - m(\beta))^2 }{S_eps(a,b) 
= OT_eps(a,b) - 0.5 OT_eps(a,a) - 0.5 OT_eps(b,b) + 0.5*eps (m(a) - m(b))²}

with transport cost \eqn{OT_\varepsilon(\alpha,\beta)}{OT_eps(a,b)}, regularization parameter \eqn{\varepsilon}{eps}
and supply and demand mass \eqn{m(\alpha)}{m(a)} and \eqn{m(\beta)}{m(b)}.

The transport costs are calculated using the Sinkhorn algorithm \code{\link[unbalancedTransport]{regularizedTransport}}.


\insertRef{Sejourne2019}{unbalancedTransport}
}
\examples{
I <- 1000
J <- 1000
X <- seq(0,1,length.out = I)
Y <- seq(0,1,length.out = J)
p <- supplyExample
q <- demandExample

supply <- list(p,X)
demand <- list(q,Y)

eps <- 1e-3 
supplyDiv <- list("KL", 0.04)
demandDiv <- list("KL", 0.04)

sinkhorn_divergence(supply, demand, supplyDiv, demandDiv, eps, p = 2)


}
