% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmSinkhorn.R
\name{sinkhornAlgorithm}
\alias{sinkhornAlgorithm}
\title{The Sinkhorn Algorithm}
\usage{
sinkhornAlgorithm(
  supplyList,
  demandList,
  maxIteration,
  eps,
  tol = 1e-05,
  method = "euclidean",
  exp = 1,
  p = 2,
  wfr = FALSE,
  costMatrix = NULL
)
}
\arguments{
\item{supplyList}{A list containing the information about the supply measure, 
divergence function and parameter. The first element is the supply measure itself,
the second the abbreviation of the divergence function ("KL" for Kullback Leibner, 
"TV" for total variation, "RG" for range constraint, "Power", "Berg" and "Hellinger") followed by the parameters 
needed for the divergence function: 

In case of the KL divergence \eqn{F_1 = \lambda \cdot KL()}{F1 = lambda * KL()} and the 
regularization parameter \eqn{\lambda}{lambda} has to be provided.

The same structure is used for the TV, Berg and Hellinger divergence. Only the parameter for lambda has to be provided.

The Power divergence also needes the conjugate exponent as additional parameter.

The divergence function associated with the range constraint needs two parameters, that define the upper and 
lower bound. 



If the cost matrix is not provided, the support of the supply measure has to be provided as last element in the list. This can
be omitted if a cost matrix is passed as argument.}

\item{demandList}{A list containing the information about the demand measure. It has to have the same structure as the supplyList.}

\item{maxIteration}{(optional) The maximum number of iterations.}

\item{eps}{A positive numeric value for epsilon.}

\item{tol}{(optional) A numeric value. If the change of the dual variables from one step to the next is smaller than this value. The algorithm
is terminated as it is already converged. The default value is 1e-5.}

\item{method}{(optional) Determines the method that is used to compute the cost matrix.
\itemize{
\item "euclidean"
\item "minkowski"
\item "maximum" 
}
The default value is "euclidean".}

\item{exp}{(optional) The exponent that is applied to the cost matrix. Can be used to compute quadratic cost. The default value is 1.}

\item{p}{(optional) Parameter for the minkwski cost function. Can be omitted if either "euclidean" or "maximum" is used. The default value is 2.}

\item{wfr}{(optional) Computes the cost matrix needed for the Wasserstein-Fisher-Rao distance \eqn{c(x,y) = -\log(\cos^2_+(d(x,y)))}{c(x,y) = -log(cos_+(d(x,y)²))}.
The default value is "false".}

\item{costMatrix}{(optional) Instead of having the algorithm compute the cost matrix, a custom cost matrix can be passed to the algorithm.}
}
\description{
This functions uses to Sinkhorn algorithm to solve the regularized unbalanced optimal transport problem.
}
\details{
The algorithm minimizes the regularized unbalanced optimal transport problem given by

\eqn{\min_{r \in \!R^{X \times Y}_+} \cdot F_1(P_X r) + \cdot F_2(P_Y r) + \varepsilon KL(r|K)}{min_r F1(P_X r) +  F_2(P_Y r) + epsilon * KL(r|K)}

where F_X and F_Y are divergence functions, \eqn{P_X r} and \eqn{P_Y r} the marginals of \eqn{r}, 
\eqn{KL(\cdot | \cdot)} the Kullback Leibner divergence and \eqn{K = \exp(-c(x,y)/eps)}{K = exp(-c(x,y)/eps)} the kernel associated with the cost matrix.


Four divergence functions are available to use with this algorithm: 
\itemize{
  \item Kullback-Leibner divergence ("KL")
  \item Total variation ("TV")
  \item The divergence associated with the range constraint ("RG")
  \item The Power divergence ("Power")
}

Two special cases of the Power divergence are directly available: 
The Berg divergence ("Berg") uses the conjugate exponent \eqn{r = -1} and the Hellinger divergence ("Hellinger") with \eqn{r = 0}.
 

Instead of providing a cost matrix to the algorithm, it is possible to provide the support of the supply and demand measures and 
have the function compute the cost matrix. Four additional parameters can be provided to define the cost matrix to compute. 


INSERT REF
}
