% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StabilizedScaling_Rcpp}
\alias{StabilizedScaling_Rcpp}
\title{The stabilized Scaling Algorithm}
\usage{
StabilizedScaling_Rcpp(
  costMatrix,
  supply,
  demand,
  lambdaSupply,
  alphaSupply,
  betaSupply,
  lambdaDemand,
  alphaDemand,
  betaDemand,
  DivSupply,
  DivDemand,
  iterMax,
  epsvec,
  tol = 1e-10
)
}
\arguments{
\item{costMatrix}{A numeric matrix.}

\item{supply}{A numeric vector}

\item{demand}{A numeric vector}

\item{lambdaSupply}{Parameter for the supply proxdiv function}

\item{alphaSupply}{numeric Value}

\item{betaSupply}{numeric value}

\item{lambdaDemand}{Parameter for the demand proxdiv function}

\item{alphaDemand}{numeric Value}

\item{betaDemand}{numeric value}

\item{DivSupply}{Parameter indicating the divergence function to use for the supply proxdiv function}

\item{DivDemand}{Parameter indicating the divergence function to use for the demand proxdiv function}

\item{iterMax}{Maximum number of iterations}

\item{epsvec}{A numeric vector of decreasing epsilon values.}

\item{tol}{num vale}
}
\value{
The optimal transport plan
}
\description{
C++ implementation of the log-domain stabilized Version of the Scaling
Algorithm.
}
