% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmTree.R
\name{treeAlgorithm}
\alias{treeAlgorithm}
\title{Tree Metric Unbalanced Transport}
\usage{
treeAlgorithm(
  tree,
  supply,
  demand,
  creationCost,
  destructionCost,
  output = "cost"
)
}
\arguments{
\item{tree}{A tree in list format. The first element holds the index of the
root node. Every other element holds a vector with three values describing an
edge: The parent node, the child node and the edges weight.}

\item{supply}{A numeric supply vector}

\item{demand}{A numeric demand vector}

\item{creationCost}{A creation cost vector}

\item{destructionCost}{A destruction cost vector}

\item{output}{Either "list", "transportPlan", "cost", or "both".}
}
\value{
the transport cost, import and export vector, transport list or plan if specified in 'output'
}
\description{
Tree Metric Unbalanced Transport
}
\examples{

tree <- list(1, c(1, 2, 1), c(2, 3, 1), c(3, 4, 1), c(3, 5, 1),
c(2, 6, 1), c(1, 7, 1), c(7, 8, 1), c(7, 9, 1),
c(9, 10, 1), c(9, 11, 1), c(11, 12, 1), c(11, 13, 1))

plotTree(tree)

p <- rep(2,13)
q <- rep(1,13)

x <- c(0,0,1,2,0,0,0,0,0,0,2,0,0)
y <- c(0,0,0,0,0,1,0,1,0,0,0,1,1)
transport <- treeAlgorithm(tree,x,y,p,q, output = "list")

plotTree(tree, tList = transport$transportList, supply = x, demand = y)

}
